# module_2-1_snake/backend/openapi.yaml
openapi: 3.1.0
info:
  title: Snake Arcade API
  version: "1.0.0"
  description: >
    API backing the Snake web client. Supports user authentication,
    leaderboard submissions, and live match watching.
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        username:
          type: string
      required: [username]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: Session token used in Authorization header
        user:
          $ref: '#/components/schemas/User'
      required: [token, user]
    LeaderboardEntry:
      type: object
      properties:
        username:
          type: string
        score:
          type: integer
          minimum: 0
      required: [username, score]
    SubmitScoreRequest:
      type: object
      properties:
        score:
          type: integer
          minimum: 0
      required: [score]
    SubmitScoreResponse:
      type: object
      properties:
        username:
          type: string
          description: Authenticated username or "guest" when unauthenticated
        score:
          type: integer
          description: Highest score recorded for this username
      required: [username, score]
    MatchSummary:
      type: object
      properties:
        id:
          type: string
        player:
          type: string
        mode:
          type: string
          enum: [pass-through, walls]
      required: [id, player, mode]
    Point:
      type: object
      properties:
        x:
          type: integer
        y:
          type: integer
      required: [x, y]
    Board:
      type: object
      properties:
        cols:
          type: integer
        rows:
          type: integer
      required: [cols, rows]
    GameFrame:
      type: object
      properties:
        id:
          type: string
          description: Match identifier
        player:
          type: string
        mode:
          type: string
          enum: [pass-through, walls]
        board:
          $ref: '#/components/schemas/Board'
        snake:
          type: array
          items:
            $ref: '#/components/schemas/Point'
        food:
          $ref: '#/components/schemas/Point'
        score:
          type: integer
        state:
          type: string
          enum: [running, dead]
      required: [id, player, mode, board, snake, food, score, state]
paths:
  /auth/signup:
    post:
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required: [username, password]
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Username already exists or validation error
  /auth/login:
    post:
      summary: Authenticate and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required: [username, password]
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /auth/logout:
    post:
      summary: End the current session
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logged out
        '401':
          description: Missing or invalid token
  /auth/me:
    get:
      summary: Get the current authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Missing or invalid token
  /leaderboard:
    get:
      summary: Retrieve the top leaderboard entries
      responses:
        '200':
          description: Sorted leaderboard (highest score first, max 10 entries)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'
  /scores:
    post:
      summary: Submit a score for the active session
      description: >
        Authenticated requests attribute the score to the logged-in user.
        Unauthenticated requests are recorded under the "guest" user.
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitScoreRequest'
      responses:
        '200':
          description: Score recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitScoreResponse'
        '400':
          description: Invalid score payload
  /watch/matches:
    get:
      summary: List live matches available for watching
      responses:
        '200':
          description: Live matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchSummary'
  /watch/{matchId}/stream:
    get:
      summary: Stream live match frames (Server-Sent Events)
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event stream of match frames
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                frame:
                  summary: Example SSE event payload
                  value: |
                    event: frame
                    data: {"id":"arena-1","player":"spectre","mode":"pass-through","board":{"cols":10,"rows":10},"snake":[{"x":5,"y":5}],"food":{"x":3,"y":7},"score":40,"state":"running"}
        '404':
          description: Match not found